Add and use expect* methods more + some cleanup
This commit uses the expectTrait methods and hasTrait methods more
instead of using .get(). This will give much more clear error messages
in the event of a logical error in code.

This also adds expect* methods that expect a shape to be of a specific
type. This cleans up a lot of call sites that were using an unsafe
.get() after calling .asBlob, asString, etc.